+++ r4/src/test/java/io/github/jy95/fds/r4/translators/MaxDosePerPeriodTest.java
@@ -1,33 +1,21 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractMaxDosePerPeriodTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Quantity;
 import org.hl7.fhir.r4.model.Ratio;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+public class MaxDosePerPeriodTest extends AbstractMaxDosePerPeriodTest<FDSConfigR4, Dosage> {
 
-public class MaxDosePerPeriodTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoMaxDosePerPeriod(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.MAX_DOSE_PER_PERIOD);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithMaxDosePerPeriod(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateWithMaxDosePerPeriod() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity numerator = new Quantity(10);
@@ -38,21 +26,24 @@ public class MaxDosePerPeriodTest extends AbstractFhirTest {
         denominator.setSystem("http://hl7.org/fhir/ValueSet/units-of-time");
         ratio.setDenominator(denominator);
         dosage.setMaxDosePerPeriod(ratio);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.MAX_DOSE_PER_PERIOD);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    private String getExpectedText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "up to a maximum of 10 mg per day";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "jusqu’à un maximum de 10 mg par jour";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "bis zu einer maximalen Menge von 10 mg pro Tag";
-        } else {
-            return "tot een maximum van 10 mg per dag";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
+    }
+
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
+    }
+
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
